chore: circular dep resolution and dropped redundant types
